refactor: move value package up to tsdb
